mx6pep: disable STOP_LIGHT platform idle state
Transitions to STOP_LIGHT were causing a bugcheck because when the
PEP transitioned the GPT timer to off, the system had no always-on
counter to replace it with. The always-on counter was
provided by the RTC, but RTC support was removed with the recent fix
to the HAL timer extension. It could be added back in, but the risks
of having STOP_LIGHT run only in UpdateOS are not worth the reward.
STOP_LIGHT mode only activates in UpdateOS
because it is normally blocked by USB. This code path does not
get enough regular coverage to justify having it on in UpdateOS
where it is difficult to debug.