Add provisioning documentation and functionality
This patch adds documentation and support for the reference
provisioning flow along with an updated configuration to support it on
the HummingBoardEdge i.MX6 Quad.

Common.mk now stores the SRKH value in the SPL Device Tree for
automatically enabling High Assurance Boot.

provision.py is a simple reference script for a provisioning host that
demonstrates the remote endpoint that the provisioning firmware expects.